requirements.txt
fastapi
uvicorn
python-dotenv
cohere
pydantic
python-multipart
requests
openai
qdrant-client